requirements.txt
Flask==2.3.2
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-api-python-client==2.90.0
requests==2.31.0
Werkzeug==2.3.7